refactor: splite technicalInterview component to smaller and extract custom hooks
